requirements.txt
Django>=5.0
psycopg2-binary>=2.9.9
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
joblib>=1.3

🛠️ POSTGRESQL SETUP (REQUIRED - before Django setup)
1. Install: https://www.postgresql.org/download/windows/
2. Create DB: CREATE DATABASE dvdrental; (user: postgres)
3. Download & restore sample DB: https://github.com/postgres-samples/dvdrental/releases
   - pgAdmin: Right-click dvdrental > Restore > dvdrental.tar
   - CLI: psql -U postgres -d dvdrental -f dvdrental.sql
4. Update PASSWORD in dvd_analytics/settings.py if needed

Full guide: README.md

🔧 OTHER USE CASES (Update Customer/Film/Rental)
Use Revenue as template:
1. Models: case_studies/{customer|film|rental}/models.py
2. Analytics: create */analytics.py
3. Views: */views.py + urls.py
4. Dashboard: import/call in dashboard/views.py
5. Migrations: python manage.py makemigrations case_studies.{app}
6. Add to INSTALLED_APPS in settings.py if new app

🚀 DATA PIPELINE / ETL (Revenue Analytics)

Run ETL to populate data:
$ python manage.py run_revenue_etl

Flow:
- Extract from Postgres dvdrental DB (payment chain)
- Transform → Save CSV raw_data/revenue_clean.csv
- Load to Django revenue_transactions model
- Refresh OLAP aggregates (RevenueDailyStoreAgg)

Dashboard/analytics read CSV/DB post-run.

Cron example (daily 2AM):
0 2 * * * cd /path/to/project && /path/to/python manage.py run_revenue_etl
